Move sarray subslice out of extension
- Move subslice method into unity_sarray
- Add two apply callback methods to additional_sframe_utilities
